+++ Sources/KeyboardKit/Emojis/EmojiKeyboardConfiguration.swift
@@ -26,12 +26,14 @@ public struct EmojiKeyboardConfiguration {
         itemSize: CGFloat = 40,
         font: Font = .system(size: 33),
         categoryFont: Font = .system(size: 20),
+        systemFont: Font = .system(size: 16),
         rows: Int = 5,
         horizontalSpacing: CGFloat = 10,
         verticalSpacing: CGFloat = 6) {
         self.itemSize = itemSize
         self.font = font
         self.categoryFont = categoryFont
+        self.systemFont = systemFont
         self.rows = rows
         self.horizontalSpacing = horizontalSpacing
         self.verticalSpacing = verticalSpacing
@@ -40,6 +42,7 @@ public struct EmojiKeyboardConfiguration {
     public let itemSize: CGFloat
     public let font: Font
     public let categoryFont: Font
+    public let systemFont: Font
     public let rows: Int
     public let horizontalSpacing: CGFloat
     public let verticalSpacing: CGFloat